requirements.txt
Biopython
pandas
regex
matplotlib
seaborn
pytest